requirements.txt
# Email and IMAP handling
imapclient==3.0.1
pyzmail36==1.0.5

# PDF generation
weasyprint==62.3
Jinja2==3.1.3

# reMarkable Cloud API
rmapy==0.4.2

# Configuration and utilities
PyYAML==6.0.1
python-dotenv==1.0.1

# Date handling
python-dateutil==2.8.2
